Separate provider lifecycle from event translation
